requirements_updated.txt
# Core dependencies
streamlit==1.31.0
PyMuPDF==1.23.26
python-docx==1.1.0
fpdf2==2.7.8

# Data processing
pandas==2.2.0
numpy==1.26.3

# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality
flake8==7.0.0
black==24.1.1
isort==5.13.2
mypy==1.8.0

# Security
bandit==1.7.6

# Logging and monitoring
python-json-logger==2.0.7

# Environment management
python-dotenv==1.0.1

# HTTP client for API calls
httpx==0.26.0

# OCR and Computer Vision
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0

# Medical NER and NLP
spacy>=3.6.1,<3.7.0
scispacy==0.5.3
# en_core_sci_sm model will be installed separately

# PDF processing
PyMuPDF==1.23.26
reportlab==4.0.8
weasyprint==60.2

# Data processing
pandas==2.2.0
numpy==1.26.3

# Machine learning
scikit-learn==1.4.0
joblib==1.3.2

# Clinical data processing (simplified)
# medspacy==1.0.0  # Has complex dependencies, install separately if needed
